Moved request control plugins to epp/framework/plugins
